Remove Drawwy-Shop submodule, update item descriptions, and adjust price handling in shop

+++ data/shop_updater.py
@@ -82,7 +82,7 @@ def check_for_shop_updates():
         print(all_path)
         for p in os.listdir(UPDATE_DIR):
             p=os.path.join(UPDATE_DIR, p)
-            if p not in all_path:
+            if p not in all_path and os.path.isfile(p):
                 os.remove(p)
 
         # Mettre à jour la dernière date de mise à jour
